refactor(wiki): share parsing helpers between catalogue and compare scripts
Move duplicated wiki table parsing out of internal_svgs into @internal/wiki,
which already owned the compare-page snapshot pipeline and its tests.

- Add fetchWikiPage, catalogueWikiConfigs, and importWikiCatalogue script
- Derive catalogue page slugs from countryWikiConfigs to avoid config drift
- Export shared parsing utilities from @internal/wiki index
- Keep a thin deprecated shim at script-import-wiki-catalogue.ts
